refactor: consolidate app into src
