+++ src/components/ListItem.tsx
@@ -13,7 +13,7 @@ import {
 import { lightTheme, getPriorityColor, getStatusColor } from '../theme/theme';
 import { formatDue, isOverdue, isToday } from '../utils/date';
 import { getStatusEmoji, getPriorityEmoji } from '../utils/status';
-import type { Task } from '../../src/database/types';
+import type { Task } from '../data/types';
 
 interface ListItemProps {
   task: Task;
