+++ src/components/SegmentedControl.tsx
